Make makeXXXSpliterator final
